+++ hari_client/models/models.py
@@ -945,6 +945,8 @@ class AttributeCreate(BaseModel):
     cumulated_frequency: typing.Any | None = None
     frequency: dict[str, int] | None = None
     question: str | None = None
+    repeats: int | None = None
+    possible_values: list[str | int | float | bool] | None = None
 
     @pydantic.model_validator(mode="before")
     @classmethod
@@ -987,6 +989,9 @@ class Attribute(BaseModel):
     question: str | None = None
     ml_predictions: dict[str, float] | None = None
     ml_probability_distributions: dict[str, float] | None = None
+    cant_solve_ratio: float | None = None
+    repeats: int | None = None
+    possible_values: list[str | int | float | bool] | None = None
 
 
 class AttributeResponse(BaseModel):
@@ -1043,6 +1048,20 @@ class AttributeResponse(BaseModel):
         description="A point estimate for the probability associated with each category"
         ", obtained from the full Dirichlet distribution predicted by the"
         " model.",
+    )
+    cant_solve_ratio: float | None = pydantic.Field(
+        default=None,
+        title="Can't Solve Ratio",
+    )
+    repeats: int | None = pydantic.Field(
+        default=None,
+        title="Repeats",
+        description="Number of repeats for this attribute",
+    )
+    possible_values: list[str | int | float | bool] | None = pydantic.Field(
+        default=None,
+        title="Possible Values",
+        description="Possible values for this attribute",
     )